Package classes into their respective packages.
