--- src/main/java/Command.java
@@ -1,7 +1,0 @@
-abstract class Command {
-    abstract void execute(TaskList tasks, Ui ui, Storage storage) throws DukeException;
-
-    public boolean isExit() {
-        return false;
-    }
-}
