+++ src/main/java/Duke/Commands/CommandExit.java
@@ -1,4 +1,10 @@
-public class CommandExit extends Command{
+package Duke.Commands;
+
+import Duke.System.Storage;
+import Duke.System.TaskList;
+import Duke.System.Ui;
+
+public class CommandExit extends Command {
     @Override
     public void execute(TaskList tasks, Ui ui, Storage storage) {
         ui.showExit();
