+++ src/main/java/Duke/Commands/CommandHelp.java
@@ -1,3 +1,10 @@
+package Duke.Commands;
+
+import Duke.DukeException.DukeException;
+import Duke.System.Storage;
+import Duke.System.TaskList;
+import Duke.System.Ui;
+
 public class CommandHelp extends Command {
     @Override
     public void execute(TaskList tasks, Ui ui, Storage storage) throws DukeException {
