+++ src/main/java/Duke/Commands/CommandToDo.java
@@ -1,3 +1,12 @@
+package Duke.Commands;
+
+import Duke.DukeException.DukeException;
+import Duke.System.Storage;
+import Duke.System.TaskList;
+import Duke.System.Ui;
+import Duke.Tasks.Task;
+import Duke.Tasks.TaskToDo;
+
 public class CommandToDo extends Command {
     private final String name;
