+++ src/main/java/Duke/DukeException/DukeException.java
@@ -1,3 +1,5 @@
+package Duke.DukeException;
+
 public class DukeException extends Exception{
     public DukeException(String msg) {
         super(msg);
