+++ src/main/java/Duke/Enums/Commands.java
@@ -1,3 +1,5 @@
+package Duke.Enums;
+
 public enum Commands {
     TODO, EVENT, DEADLINE, LIST, MARK, UNMARK, DELETE, EXIT, HELP;
 }
