+++ src/main/java/Duke/Main/Duke.java
@@ -1,3 +1,13 @@
+package Duke.Main;
+
+import Duke.Commands.Command;
+import Duke.Commands.CommandEmpty;
+import Duke.DukeException.DukeException;
+import Duke.System.Parser;
+import Duke.System.Storage;
+import Duke.System.TaskList;
+import Duke.System.Ui;
+
 public class Duke {
     private final Storage storage;
     private TaskList tasks;
