+++ src/main/java/Duke/Tasks/TaskDeadline.java
@@ -1,15 +1,20 @@
-public class TaskEvent extends Task {
+package Duke.Tasks;
+
+import Duke.Time.ManagerDate;
+import Duke.Time.ManagerTime;
+
+public class TaskDeadline extends Task {
     private final String date;
     private final String time;
 
-    TaskEvent(String name, String date, String time) {
+    public TaskDeadline(String name, String date, String time) {
         super(name);
         this.date = date;
         this.time = new ManagerTime(time).getFormat24();
     }
 
     public String getPrefix() {
-        return "E";
+        return "D";
     }
 
     @Override
@@ -24,7 +29,7 @@ public class TaskEvent extends Task {
 
     @Override
     public String toString() {
-        String prefix = "[E]";
+        String prefix = "[D]";
         return prefix + super.toString() + " on: "
                 + ManagerDate.formatDate(this.date) + " "
                 + this.time;
