+++ src/main/java/Duke/Tasks/TaskEvent.java
@@ -1,15 +1,20 @@
-public class TaskDeadline extends Task {
+package Duke.Tasks;
+
+import Duke.Time.ManagerDate;
+import Duke.Time.ManagerTime;
+
+public class TaskEvent extends Task {
     private final String date;
     private final String time;
 
-    TaskDeadline(String name, String date, String time) {
+    public TaskEvent(String name, String date, String time) {
         super(name);
         this.date = date;
         this.time = new ManagerTime(time).getFormat24();
     }
 
     public String getPrefix() {
-        return "D";
+        return "E";
     }
 
     @Override
@@ -24,7 +29,7 @@ public class TaskDeadline extends Task {
 
     @Override
     public String toString() {
-        String prefix = "[D]";
+        String prefix = "[E]";
         return prefix + super.toString() + " on: "
                 + ManagerDate.formatDate(this.date) + " "
                 + this.time;
